replace query plan types entirely
